+++ python/qibuild/actions/install.py
@@ -18,8 +18,6 @@ def configure_parser(parser):
     group.add_argument("--prefix", metavar="PREFIX",
         help="value of CMAKE_INSTALL_PREFIX, defaults to '/'")
     group.add_argument("destdir", metavar="DESTDIR")
-    group.add_argument("--runtime", action="store_true",
-        help="install runtime components only")
     group.add_argument("--include-deps", action="store_true", dest="include_deps",
         help="Include dependencies when installing (this is the default)")
     group.add_argument("--no-include-deps", action="store_false", dest="include_deps",
@@ -37,37 +35,36 @@ def do(args):
     dest = os.path.join(destdir, prefix)
 
     # Resolve deps:
-    (project_names, package_names, _) = toc.resolve_deps(runtime=args.runtime)
+    (packages, projects) = qibuild.cmdparse.deps_from_args(toc, args)
 
     if toc.active_config:
         ui.info(ui.green, "Active configuration: ",
                 ui.blue, "%s (%s)" % (toc.active_config, toc.build_type))
 
     ui.info(ui.green, "The following projects")
-    for project_name in project_names:
-        ui.info(ui.green, " *", ui.blue, project_name)
-    if args.include_deps and package_names:
+    for project in projects:
+        ui.info(ui.green, " *", ui.blue, project.name)
+    if args.include_deps and packages:
         ui.info(ui.green, "and the following packages")
-        for package_name in package_names:
-            ui.info(" *", ui.blue, package_name)
+        for package in packages:
+            ui.info(" *", ui.blue, package.name)
     ui.info(ui.green, "will be installed to", ui.blue, dest)
     if args.runtime:
         ui.info(ui.green, "(runtime components only)")
 
     # Install packages to destdir:
-    if args.include_deps and package_names:
+    if args.include_deps and packages:
         print
         ui.info(ui.green, ":: ", "Installing packages")
-        for (i, package_name) in enumerate(package_names):
-            ui.info(ui.green, "*", ui.reset, "(%i/%i)" % (i+1, len(package_names)),
-                    ui.green, "Installing package", ui.blue, package_name)
+        for (i, package) in enumerate(packages):
+            ui.info(ui.green, "*", ui.reset, "(%i/%i)" % (i+1, len(packages)),
+                    ui.green, "Installing package", ui.blue, package.name)
 
-            toc.toolchain.install_package(package_name, dest, runtime=args.runtime)
+            toc.toolchain.install_package(package.name, dest, runtime=args.runtime)
         print
 
     # Install projects to destdir:
     ui.info(ui.green, ":: ", "Installing projects")
-    projects = [toc.get_project(name) for name in project_names]
     for (i, project) in enumerate(projects):
         ui.info(ui.green, "*", ui.reset, "(%i/%i)" % (i+1, len(projects)),
                 ui.green, "Installing project", ui.blue, project.name)