+++ python/qibuild/actions/package.py
@@ -9,9 +9,7 @@ import sys
 from qibuild import ui
 import qibuild
 
-def get_package_name(project,
-    version=None,
-    config=None):
+def get_package_name(project, version=None, config=None):
     """Get the package name of a project.
 
     Recognized args are:
@@ -91,13 +89,8 @@ def do(args):
     """Main entry point"""
     toc = qibuild.toc_open(args.worktree, args)
     config = toc.active_config
-    if not args.project:
-        project_name = qibuild.project.project_from_cwd()
-    else:
-        project_name = args.project
-    project = toc.get_project(project_name)
-    package_name = get_package_name(project,
-        version=args.version, config=config)
+    project = qibuild.cmdparse.project_from_args(toc, args)
+    package_name = get_package_name(project, version=args.version, config=config)
     destdir = os.path.join(toc.worktree.root, "package")
     destdir = os.path.join(destdir, package_name)
 
@@ -106,19 +99,19 @@ def do(args):
 
     if sys.platform.startswith("win") and not args.runtime:
         # Ignore the --release flag and always build in debug and in release:
-        _do_package(args, project_name, destdir, debug=True)
-        _do_package(args, project_name, destdir, debug=False)
+        _do_package(args, project.name, destdir, debug=True)
+        _do_package(args, project.name, destdir, debug=False)
     else:
         ui.info(ui.green, "> Configuring ...")
-        qibuild.run_action("qibuild.actions.configure", [project_name, "--no-clean-first"],
+        qibuild.run_action("qibuild.actions.configure", [project.name, "--no-clean-first"],
             forward_args=args)
         print
         ui.info(ui.green, "> Building ...")
-        qibuild.run_action("qibuild.actions.make", [project_name, "--no-fix-shared-libs"],
+        qibuild.run_action("qibuild.actions.make", [project.name, "--no-fix-shared-libs"],
             forward_args=args)
         print
         ui.info(ui.green, "> Installing ...")
-        qibuild.run_action("qibuild.actions.install", [project_name, destdir],
+        qibuild.run_action("qibuild.actions.install", [project.name, destdir],
             forward_args=args)
         print