+++ python/qibuild/parsers.py
@@ -77,9 +77,17 @@ def build_parser(parser):
 def project_parser(parser):
     """ Parser settings for every action using several toc projects
     """
-    parser.add_argument("-a", "--all", action="store_true",
+    group = parser.add_argument_group("dependency resolution options")
+    group.add_argument("-a", "--all", action="store_true",
         help="Work on all projects")
-    parser.add_argument("-s", "--single", action="store_true",
+    group.add_argument("-s", "--single", action="store_true",
         help="Work on specified projects without taking dependencies into account.")
+    group.add_argument("--build-deps", action="store_true",
+        help="Work on specified projects by ignoring the runtime deps. "
+             "Useful when you have lots of runtime plugins you don't want to compile "
+             "for instance")
+    group.add_argument("--runtime", action="store_true",
+        help="Work on specified projects by using only the runtime deps. "
+             "Mostly used by qibuild install --runtime")
     parser.add_argument("projects", nargs="*", metavar="PROJECT", help="Project name(s)")
-    parser.set_defaults(single=False, projects = list())
+    parser.set_defaults(single=False, build_deps=False, projects = list())